no more -Wconversion
commit 6cfb6074e2e0b42d5f278696ecbb0349398faaed
Author: Luke Videckis <lukevideckis@gmail.com>
Date:   Fri Jan 12 14:05:15 2024 -0600

    this is bugging me

commit b179236d4af8fb5315c1d8864e76768a8dd1f107
Author: Luke Videckis <lukevideckis@gmail.com>
Date:   Fri Jan 12 14:03:54 2024 -0600

    update readme

commit 959f752be4c675352d42fa2f467973b001d1a661
Author: Luke Videckis <lukevideckis@gmail.com>
Date:   Fri Jan 12 14:03:07 2024 -0600

    now it shouldnt fail

commit c620fde999285dfb3942fced180e7d178607ebb6
Author: Luke Videckis <lukevideckis@gmail.com>
Date:   Fri Jan 12 14:02:34 2024 -0600

    asdf

commit 3b22d8d13976f18721593d1a8da9d64f0397c6cf
Author: Luke Videckis <lukevideckis@gmail.com>
Date:   Fri Jan 12 14:00:33 2024 -0600

    Revert "no more -Wconversion"

    This reverts commit db6af090b7e2467d700102d09e3d23f418215cb5.

commit db6af090b7e2467d700102d09e3d23f418215cb5
Author: Luke Videckis <lukevideckis@gmail.com>
Date:   Fri Jan 12 13:57:24 2024 -0600

    no more -Wconversion

commit bc40fd7864a9c60bcc3519d93d017b2deaf60416
Author: Luke Videckis <lukevideckis@gmail.com>
Date:   Fri Jan 12 13:55:16 2024 -0600

    no more int cast nor -Wconversion checking

+++ library/convolution/and_convolution.hpp
@@ -19,7 +19,7 @@ vector<int> and_convolution(vector<int> a, vector<int> b) {
             }
     vector<int> c(1 << n);
     for (int i = 0; i < (1 << n); i++)
-        c[i] = int(1LL * a[i] * b[i] % mod);
+        c[i] = 1LL * a[i] * b[i] % mod;
     for (int i = 0; i < n; i++)
         for (int j = 0; j < (1 << n); j++)
             if (!((j >> i) & 1))
