+++ library/convolution/lcm_convolution.hpp
@@ -17,7 +17,7 @@ vector<int> lcm_convolution(const vector<int>& a, const vector<int>& b) {
         for (int j = i; j < n; j += i)
             sum_a[j] += a[i], sum_b[j] += b[i];
         sum_a[i] %= mod, sum_b[i] %= mod;
-        c[i] = int((c[i] + sum_a[i] * sum_b[i]) % mod);
+        c[i] = (c[i] + sum_a[i] * sum_b[i]) % mod;
         for (int j = i + i; j < n; j += i)
             if ((c[j] -= c[i]) < 0) c[j] += mod;
     }
